updated sharshipsPage
